requirements/base.txt
python-slugify==8.0.4
pillow==11.3.0
python-magic==0.4.27
whitenoise==6.9.0
django-phonenumber-field[phonenumbers]==8.1.0
babel==2.17.0  # The SplitPhoneNumberField requires the babel package.

# Payments
django-payments[stripe] == 3.1.0

# Django
# ------------------------------------------------------------------------------
Django==5.2.5
djangorestframework==3.16.1
django-allauth[socialaccount]==65.10.0
